Fixes#21 To improve resource creation/deletion for deployment flow
Current deployment flow deletes the deployments from the stage, run
namesapces because the way it processes templates. Which in turn
disrupts running deployments for a while and triggers deployments
twice.
This patch fixes the way it processes the templates and creates
the resources. Now the deployment flow process the template,
creates ImageSream, triggers builds and deploys the application
resources. It doesn't delete the resources from
any namespaces. It doesn't triggers the deployment if given
the right DC triggers.

Fixes:
	https://github.com/fabric8io/osio-pipeline/issues/21
	https://openshift.io/openshiftio/Openshift_io/plan/detail/341